+++ src/Request/People.php
@@ -1509,7 +1509,7 @@ class People extends RequestCollection
     public function getSharePrefill()
     {
         return $this->ig->request('banyan/banyan/')
-            ->addParam('views', '["story_share_sheet","threads_people_picker","reshare_share_sheet"]')
+            ->addParam('views', '["story_share_sheet","direct_user_search_nullstate","forwarding_recipient_sheet","threads_people_picker","direct_inbox_active_now","group_stories_share_sheet","reshare_share_sheet","direct_user_search_keypressed"]')
             ->getResponse(new Response\SharePrefillResponse());
     }